apitest/requirements.txt
Django==2.2.7
djangorestframework==3.10.3
djangorestframework-api-key==1.4.1
djangorestframework-jwt==1.11.0
django_environ==0.4.5